feat: fractional --max-hours and AWS_DEFAULT_REGION env support
- Section__Shutdown: convert float hours to integer seconds for systemd
  (--on-active=360s is unambiguous; 0.1h = 360s = 6 minutes works cleanly)
- Spec__CLI__Builder: max_hours param type int → float so typer accepts 0.1
- Spec__CLI__Defaults: DEFAULT_MAX_HOURS = 1.0; DEFAULT_REGION reads
  AWS_DEFAULT_REGION env var (falls back to eu-west-2)
- Schema__Local_Claude__Create__Request: max_hours int → float
- Local_Claude__Service: DEFAULT_REGION reads AWS_DEFAULT_REGION; create_node
  uses float() not int() so 0.1h is not truncated to 0
- Local_Claude__User_Data__Builder: max_hours int → float

+++ sg_compute_specs/local_claude/service/Local_Claude__User_Data__Builder.py
@@ -36,7 +36,7 @@ class Local_Claude__User_Data__Builder(Type_Safe):
                      max_model_len        : int   = 65536                                        ,
                      kv_cache_dtype       : str   = 'fp8'                                        ,
                      gpu_memory_utilization: float = 0.92                                        ,
-                     max_hours            : int   = 1                                            ,
+                     max_hours            : float = 1.0                                          ,
                      gpu_required         : bool  = True                                         ,
                      with_claude_code     : bool  = True                                         ,
                      with_sgit            : bool  = True                                         ,
